Various clean up changes; added lidar simulation test
