--- lidar_scanner_node/README.md
@@ -1,1 +1,0 @@
-# Simulated Lidar Scanner
