--- lidar_scanner_node/test/lidar_test.launch
@@ -1,16 +1,0 @@
-<launch>
-
-  <param name="robot_description" command="$(find xacro)/xacro --inorder '$(find lidar_scanner_node)/test/lidar_test.xacro'" />
-
-  <include file="$(find lidar_scanner_node)/launch/scanner_relocator.launch">
-    <arg name ="world_frame_name" value="odom"/>
-    <arg name ="scanner_parent_frame" value="lidar_parent"/>
-    <arg name ="scanner_params_file" value="$(find lidar_scanner_node)/test/test_params.yaml"/>
-  </include>
-
-  <node name="rviz" pkg="rviz" type="rviz" args="-d $(find lidar_scanner_node)/config/sensor_relocator.rviz"/>
-
-  <node name="joint_state_publisher" pkg="joint_state_publisher" type="joint_state_publisher"/>
-  <node name="robot_state_publisher" pkg="robot_state_publisher" type="state_publisher"/>
-
-</launch>
