+++ simulated_lidar_scanner/README.md
@@ -1,0 +1,9 @@
+# Simulated Lidar Scanner
+
+## Description
+
+The Simulated Lidar Scanner package is a ROS wrapper for
+
+## Usage
+
+
