+++ simulated_lidar_scanner/config/sensor_relocator.rviz
