+++ simulated_lidar_scanner/src/synthetic_lidar_scanner/vtkRay.cxx
@@ -16,7 +16,7 @@
  *
  *=========================================================================*/
 
-#include "lidar_scanner_node/synthetic_lidar_scanner/vtkRay.h"
+#include "simulated_lidar_scanner/synthetic_lidar_scanner/vtkRay.h"
 
 #include "vtkObjectFactory.h" //for new() macro
 #include "vtkMath.h"
